// src/shared/api/models/putApiEditSalesIdBodyData.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PutApiEditSalesIdBodyDataCustomerId } from './putApiEditSalesIdBodyDataCustomerId';
import type { PutApiEditSalesIdBodyDataSalesType } from './putApiEditSalesIdBodyDataSalesType';
import type { PutApiEditSalesIdBodyDataItemsItem } from './putApiEditSalesIdBodyDataItemsItem';
import type { PutApiEditSalesIdBodyDataSalesTaxesItem } from './putApiEditSalesIdBodyDataSalesTaxesItem';
import type { PutApiEditSalesIdBodyDataSalesDiscountsItem } from './putApiEditSalesIdBodyDataSalesDiscountsItem';
import type { PutApiEditSalesIdBodyDataPaymentItem } from './putApiEditSalesIdBodyDataPaymentItem';

export type PutApiEditSalesIdBodyData = {
    salesDate?: string;
    customerId?: PutApiEditSalesIdBodyDataCustomerId;
    customerName?: string;
    subTotal: number;
    totalTax: number;
    netTotal: number;
    tenderAmount: number;
    change: number;
    roundOff: number;
    tips?: number;
    salesNote?: string;
    salesType?: PutApiEditSalesIdBodyDataSalesType;
    sellBy: number;
    items: PutApiEditSalesIdBodyDataItemsItem[];
    salesTaxes?: PutApiEditSalesIdBodyDataSalesTaxesItem[];
    salesDiscounts?: PutApiEditSalesIdBodyDataSalesDiscountsItem[];
    payment: PutApiEditSalesIdBodyDataPaymentItem[];
};